Move Clone function to git pkg where it belongs
