--- ginclient/git.go
@@ -1,130 +1,0 @@
-package ginclient
-
-import (
-	"bytes"
-	"fmt"
-	"runtime"
-	"strings"
-
-	"github.com/G-Node/gin-cli/git"
-	"github.com/G-Node/gin-cli/util"
-)
-
-// TODO: Transfer these functions to the git subpackage. Make changes where necessary so they don't depend on the ginclient.Client type.
-
-const progcomplete = "100%"
-
-// NOTE: Duplicate function
-func cutline(b []byte) (string, bool) {
-	idx := -1
-	cridx := bytes.IndexByte(b, '\r')
-	nlidx := bytes.IndexByte(b, '\n')
-	if cridx > 0 {
-		idx = cridx
-	} else {
-		cridx = len(b) + 1
-	}
-	if nlidx > 0 && nlidx < cridx {
-		idx = nlidx
-	}
-	if idx <= 0 {
-		return string(b), true
-	}
-	return string(b[:idx]), false
-}
-
-// NOTE: Duplicate function
-func splitRepoParts(repoPath string) (repoOwner, repoName string) {
-	repoPathParts := strings.SplitN(repoPath, "/", 2)
-	repoOwner = repoPathParts[0]
-	repoName = repoPathParts[1]
-	return
-}
-
-// Clone downloads a repository and sets the remote fetch and push urls.
-// Setting the Workingdir package global affects the working directory in which the command is executed.
-// The status channel 'clonechan' is closed when this function returns.
-// (git clone ...)
-func (gincl *Client) Clone(repoPath string, clonechan chan<- git.RepoFileStatus) {
-	fn := fmt.Sprintf("Clone(%s)", repoPath)
-	defer close(clonechan)
-	remotePath := fmt.Sprintf("ssh://%s@%s/%s", gincl.GitUser, gincl.GitHost, repoPath)
-	args := []string{"clone", "--progress", remotePath}
-	if runtime.GOOS == "windows" {
-		// force disable symlinks even if user can create them
-		// see https://git-annex.branchable.com/bugs/Symlink_support_on_Windows_10_Creators_Update_with_Developer_Mode/
-		args = append([]string{"-c", "core.symlinks=false"}, args...)
-	}
-	cmd := git.Command(args...)
-	err := cmd.Start()
-	if err != nil {
-		clonechan <- git.RepoFileStatus{Err: ginerror{UError: err.Error(), Origin: fn}}
-		return
-	}
-
-	var line string
-	var stderr []byte
-	var status git.RepoFileStatus
-	status.State = "Downloading repository"
-	clonechan <- status
-	var rerr error
-	readbuffer := make([]byte, 1024)
-	var nread, errhead int
-	var eob, eof bool
-	// git clone progress prints to stderr
-	for eof = false; !eof; nread, rerr = cmd.ErrReader.Read(readbuffer) {
-		if rerr != nil && errhead == len(stderr) {
-			eof = true
-		}
-		stderr = append(stderr, readbuffer[:nread]...)
-		for eob = false; !eob || errhead < len(stderr); line, eob = cutline(stderr[errhead:]) {
-			if len(line) == 0 {
-				errhead++
-				break
-			}
-			errhead += len(line) + 1
-			words := strings.Fields(line)
-			status.FileName = repoPath
-			if strings.HasPrefix(line, "Receiving objects") {
-				if len(words) > 2 {
-					status.Progress = words[2]
-				}
-				if len(words) > 8 {
-					rate := fmt.Sprintf("%s%s", words[7], words[8])
-					if strings.HasSuffix(rate, ",") {
-						rate = strings.TrimSuffix(rate, ",")
-					}
-					status.Rate = rate
-				}
-			}
-			clonechan <- status
-		}
-	}
-	errstring := string(stderr)
-	if err = cmd.Wait(); err != nil {
-		util.LogWrite("Error during clone command")
-		repoOwner, repoName := splitRepoParts(repoPath)
-		gerr := ginerror{UError: errstring, Origin: fn}
-		if strings.Contains(errstring, "does not exist") {
-			gerr.Description = fmt.Sprintf("Repository download failed\n"+
-				"Make sure you typed the repository path correctly\n"+
-				"Type 'gin repos %s' to see if the repository exists and if you have access to it",
-				repoOwner)
-		} else if strings.Contains(errstring, "already exists and is not an empty directory") {
-			gerr.Description = fmt.Sprintf("Repository download failed.\n"+
-				"'%s' already exists in the current directory and is not empty.", repoName)
-		} else if strings.Contains(errstring, "Host key verification failed") {
-			gerr.Description = "Server key does not match known/configured host key."
-		} else {
-			gerr.Description = fmt.Sprintf("Repository download failed. Internal git command returned: %s", errstring)
-		}
-		status.Err = gerr
-		clonechan <- status
-		// doesn't really need to break here, but let's not send the progcomplete
-		return
-	}
-	// Progress doesn't show 100% if cloning an empty repository, so let's force it
-	status.Progress = progcomplete
-	clonechan <- status
-	return
-}
